+++ tests/guile/suitbl-test-runner.scm
@@ -42,7 +42,7 @@
     (define number-of-tests
       (assoc-ref summary 'tests))
 
-    (unless (= 146 number-of-tests)
+    (unless (= 147 number-of-tests)
       (chain "Unexpected number of tests (~a), make sure all tests are executed and
 expected number of tests is up-to-date."
         (format #f _ number-of-tests)